requirements.txt
python-dotenv
Django
openai